server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "tavily-python>=0.7.17",
    "uvicorn>=0.40.0",
]
